# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sun Proactive - Volunteer Matching Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type VolunteerAvailability =
  (typeof VolunteerAvailability)[keyof typeof VolunteerAvailability];

export const VolunteerAvailability = {
  weekdays: "weekdays",
  weekends: "weekends",
  evenings: "evenings",
  flexible: "flexible",
} as const;

export interface Volunteer {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  skills: string[];
  availability: VolunteerAvailability;
  location?: string | null;
  bio?: string | null;
  totalHours: number;
  createdAt: string;
}

export type CreateVolunteerRequestAvailability =
  (typeof CreateVolunteerRequestAvailability)[keyof typeof CreateVolunteerRequestAvailability];

export const CreateVolunteerRequestAvailability = {
  weekdays: "weekdays",
  weekends: "weekends",
  evenings: "evenings",
  flexible: "flexible",
} as const;

export interface CreateVolunteerRequest {
  name: string;
  email: string;
  phone?: string | null;
  skills: string[];
  availability: CreateVolunteerRequestAvailability;
  location?: string | null;
  bio?: string | null;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  open: "open",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Task {
  id: number;
  title: string;
  description: string;
  requiredSkills: string[];
  location?: string | null;
  date?: string | null;
  /** Duration in hours */
  duration?: number | null;
  volunteersNeeded: number;
  status: TaskStatus;
  coordinatorName: string;
  createdAt: string;
}

export interface CreateTaskRequest {
  title: string;
  description: string;
  requiredSkills: string[];
  location?: string | null;
  date?: string | null;
  duration?: number | null;
  volunteersNeeded: number;
  coordinatorName: string;
}

export type UpdateTaskStatusRequestStatus =
  (typeof UpdateTaskStatusRequestStatus)[keyof typeof UpdateTaskStatusRequestStatus];

export const UpdateTaskStatusRequestStatus = {
  open: "open",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateTaskStatusRequest {
  status: UpdateTaskStatusRequestStatus;
}

export type MatchStatus = (typeof MatchStatus)[keyof typeof MatchStatus];

export const MatchStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
  completed: "completed",
} as const;

export interface Match {
  id: number;
  taskId: number;
  volunteerId: number;
  volunteerName: string;
  taskTitle: string;
  status: MatchStatus;
  /** Match score from 0-100 */
  matchScore: number;
  createdAt: string;
}

export interface VolunteerMatch {
  volunteer: Volunteer;
  /** Match score from 0-100 */
  matchScore: number;
  matchingSkills: string[];
  alreadyMatched: boolean;
}

export interface CreateMatchRequest {
  taskId: number;
  volunteerId: number;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatRequest {
  messages: ChatMessage[];
}

export interface ChatResponse {
  message: string;
  taskData?: CreateTaskRequest | null;
  isComplete: boolean;
}

export type ListTasksParams = {
  status?: ListTasksStatus;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  open: "open",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;