# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sun Proactive - Volunteer Matching Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /volunteers:
    get:
      operationId: listVolunteers
      tags: [volunteers]
      summary: List all volunteers
      responses:
        "200":
          description: List of volunteers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Volunteer"
    post:
      operationId: createVolunteer
      tags: [volunteers]
      summary: Register a new volunteer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVolunteerRequest"
      responses:
        "201":
          description: Volunteer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
  /volunteers/{id}:
    get:
      operationId: getVolunteer
      tags: [volunteers]
      summary: Get a volunteer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Volunteer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateVolunteer
      tags: [volunteers]
      summary: Update a volunteer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVolunteerRequest"
      responses:
        "200":
          description: Volunteer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Volunteer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        availability:
          type: string
          enum: [weekdays, weekends, evenings, flexible]
        location:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        totalHours:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - skills
        - availability
        - totalHours
        - createdAt
    CreateVolunteerRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        availability:
          type: string
          enum: [weekdays, weekends, evenings, flexible]
        location:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
      required:
        - name
        - email
        - skills
        - availability
...